+++ src/network/core/CMakeLists.txt
@@ -4,7 +4,8 @@ add_files(
     config.h
     core.cpp
     core.h
-    game.h
+    game_info.cpp
+    game_info.h
     host.cpp
     host.h
     os_abstraction.h
