+++ src/network/core/core.h
@@ -71,8 +71,6 @@ public:
 	 */
 	void Reopen() { this->has_quit = false; }
 
-	void SendGRFIdentifier(Packet *p, const GRFIdentifier *grf);
-	void ReceiveGRFIdentifier(Packet *p, GRFIdentifier *grf);
 	void SendCompanyInformation(Packet *p, const struct Company *c, const struct NetworkCompanyStats *stats, uint max_len = NETWORK_COMPANY_NAME_LENGTH);
 };
