+++ src/network/core/tcp_http.cpp
@@ -13,6 +13,7 @@
 #include "../../debug.h"
 #include "../../rev.h"
 #include "../network_func.h"
+#include "game_info.h"
 
 #include "tcp_http.h"
