+++ src/network/network_gui.h
@@ -11,6 +11,7 @@
 #define NETWORK_GUI_H
 
 #include "../company_type.h"
+#include "../date_type.h"
 #include "../economy_type.h"
 #include "../window_type.h"
 #include "network_type.h"
