+++ src/network/network_type.h
@@ -10,7 +10,7 @@
 #ifndef NETWORK_TYPE_H
 #define NETWORK_TYPE_H
 
-#include "core/game.h"
+#include "core/config.h"
 
 /** How many clients can we have */
 static const uint MAX_CLIENTS = 255;
